add test

+++ light-client/src/header/validator_set.rs
@@ -15,7 +15,7 @@ impl ValidatorSet {
     /// https://github.com/NathanBSC/bsc/blob/a910033bc52013d96ecefd8d5224d70d288c1309/consensus/parlia/snapshot.go#L226
     pub fn checkpoint(&self, turn_term: u8) -> u64 {
         let validator_size = self.validators.len() as u64;
-        return (validator_size / 2) * turn_term + 1
+        return (validator_size / 2) * turn_term as u64 + 1
     }
 }
